# tenantMicroserviceAssignment/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The Tenant Microservice APIs",
        "title": "Tenant Microservice APIs",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/tenant": {
            "get": {
                "description": "Designed to be used with some sort of admin web panel, etc.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenant"
                ],
                "summary": "This API can be used to get all tenants.",
                "responses": {
                    "200": {
                        "description": "when the operation is successful",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Tenant"
                            }
                        }
                    },
                    "500": {
                        "description": "when failed reading the database",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Adds a new tenant in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenant"
                ],
                "summary": "This API can be used to create a new tenant.",
                "parameters": [
                    {
                        "description": "create tenant request",
                        "name": "default",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateTenantRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "500": {
                        "description": "when there's some error adding the tenant, can be a database failure or a tenant already exist with the id",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tenant/credentials": {
            "put": {
                "description": "Updates a tenant credentials in the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenant"
                ],
                "summary": "This API can be used to update a tenant credentials.",
                "parameters": [
                    {
                        "description": "update tenant credentials request",
                        "name": "default",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateTenantCredentialsRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "406": {
                        "description": "when there's some error updating the tenant credentials, can be a database failure or no tenant found with the id",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tenant/license": {
            "put": {
                "description": "Updates a tenant license in the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenant"
                ],
                "summary": "This API can be used to update a tenant license.",
                "parameters": [
                    {
                        "description": "update tenant license request",
                        "name": "default",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateTenantLicenseRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "406": {
                        "description": "when there's some error updating the tenant, can be a database failure or no tenant found with the id",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tenant/status": {
            "put": {
                "description": "Updates a tenant status in the database.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenant"
                ],
                "summary": "This API can be used to update a tenant status.",
                "parameters": [
                    {
                        "description": "update tenant status request",
                        "name": "default",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UpdateTenantStatusRequestBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "406": {
                        "description": "when there's some error updating the tenant status, can be a database failure or no tenant found with the id",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/tenant/{tenant_id}": {
            "get": {
                "description": "Retrieves a tenant from the database by id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tenant"
                ],
                "summary": "This API can be used to get a tenant by id.",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tenant ID",
                        "name": "tenant_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "when the operation is successful",
                        "schema": {
                            "$ref": "#/definitions/models.Tenant"
                        }
                    },
                    "404": {
                        "description": "when no tenant is found against the provided tenant id",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.CreateTenantRequest": {
            "type": "object",
            "properties": {
                "first_name": {
                    "type": "string"
                },
                "is_enabled": {
                    "type": "string"
                },
                "license_end_date": {
                    "type": "string"
                },
                "license_start_date": {
                    "type": "string"
                },
                "license_type": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "total_user_count": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "models.Tenant": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "is_enabled": {
                    "type": "string"
                },
                "license_end_date": {
                    "type": "string"
                },
                "license_start_date": {
                    "type": "string"
                },
                "license_type": {
                    "type": "string"
                },
                "total_user_count": {
                    "type": "string"
                }
            }
        },
        "models.UpdateTenantCredentialsRequestBody": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string"
                }
            }
        },
        "models.UpdateTenantLicenseRequestBody": {
            "type": "object",
            "properties": {
                "new_license_end_date": {
                    "type": "string"
                },
                "new_license_start_date": {
                    "type": "string"
                },
                "new_license_type": {
                    "type": "string"
                }
            }
        },
        "models.UpdateTenantStatusRequestBody": {
            "type": "object",
            "properties": {
                "is_enabled": {
                    "type": "string"
                }
            }
        }
    }
}